requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
distro==1.9.0
duckduckgo_search==8.1.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.10
langchain-core==1.2.17
langchain-groq==1.1.2
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.13
lxml==6.0.2
MarkupSafe==3.0.3
multitasking==0.0.12
narwhals==2.17.0
numpy==2.4.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
peewee==4.0.1
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
praw==7.8.1
prawcore==2.4.0
primp==1.1.2
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
urllib3==2.6.3
uuid_utils==0.14.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yfinance==1.2.0
zstandard==0.25.0
